% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{B1}
\alias{B1}
\title{B1}
\usage{
B1(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo} or \code{treeshape}.}
}
\value{
An object of class \code{numeric}.
}
\description{
This calculates B1 for a phylogenetic tree. For each internal node \deqn{j}, we consider the leaf with the largest topological distance \deqn{Z_j}. Then,
\deqn{ B_1 = \sum_{j \in \mathcal{I}-root} {Z_j^{-1} }}.
}
\examples{
N=30
tree<-rtreeshape(1,tip.number=N,model="pda")[[1]]
B1(tree)
}
\references{
Pompei S, Loreto V, Tria F (2012) Phylogenetic Properties of RNA Viruses. PLoS ONE 7(9): e44849. doi:10.1371/journal.pone.0044849
Shao K, Sokal R (1990) Tree balance. Systematic Zooology 39: 266-276
}
\keyword{asymmetry}
\keyword{imbalance,}

