% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSimTree.R
\name{getCoalescenceEvents}
\alias{getCoalescenceEvents}
\title{getCoalescenceEvents}
\usage{
getCoalescenceEvents(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo}.}
}
\value{
A object class \code{numeric} giving the coalescence times.
}
\description{
Gets the times of coalescence events (internal nodes), backwards in time from most recently sampled tip.
}

