% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSimTree.R
\name{getTipTimes}
\alias{getTipTimes}
\title{getTipTimes}
\usage{
getTipTimes(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo}.}
}
\value{
A object class \code{numeric} giving the tree tip sampling times.
}
\description{
Gets the times of the tips of a heterochronously sampled tree with most recently sampled tip set to 0.
}

