% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{Ic}
\alias{Ic}
\title{Ic}
\usage{
Ic(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo} or \code{treeshape}.}
}
\value{
An object of class \code{numeric}.
}
\description{
This calculates Ic for a phylogenetic tree. This asymmetry metric uses a weighted average of the balance of internal nodes:
\deqn{ I_c = \frac{1}{(N-1)} \sum_{j \in \mathcal{I}} w_j \frac{\max(r_j,l_j)-m_j}{r_j+l_j-m_j-1}. }
}
\examples{
N=30
tree<-rtreeshape(1,tip.number=N,model="pda")[[1]]
Ic(tree)
}
\references{
\itemize{
 \item Pompei S, Loreto V, Tria F (2012) Phylogenetic Properties of RNA Viruses. PLoS ONE 7(9): e44849. doi:10.1371/journal.pone.0044849
 \item Purvis A, Agapow PM (2002) Phylogeny imbalance: Taxonomic level matters. Systematic Biology 51: 844?854.
 \item Fusco G, Cronk Q (1995) A new method for evaluating the shape of large phylogenies. Journal of Theoretical Biology 175: 235?243. doi: 10.1006/jtbi.1995.0136
}
}
\keyword{asymmetry}
\keyword{imbalance,}

