% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lbi.R
\name{lbi}
\alias{lbi}
\title{LBI}
\usage{
lbi(tree, tau = 5e-04, transform = function(x) {     x })
}
\arguments{
\item{tree}{a tree of class \code{phylo}.}

\item{tau}{a number giving the scaling factor to use, suggested by Neher et al. to be 0.0625 times the average pairwise distance in the sample.}

\item{transform}{a function by which to transform the LBI. Default is no transformation.}
}
\value{
an vector of class \code{numeric} giving the LBI for the tips and internal nodes in the same order as given in the tree \code{phylo} object.
}
\description{
This calculates the local branching index (LBI), defined by Neher, Russell and Shraiman (2014) as the the tree length surrounding a given node or tip exponentially discounted with increasing distance from that node.
}
\examples{
tree<-rtree(50)
tree.lbi<-lbi(tree)
tree$node.label<-tree.lbi[(tree$Nnode+2):length(tree.lbi)]
plot(tree, show.node.label = TRUE)
}
\references{
Neher RA, Russell CA, Shraiman BI (2014) Predicting evolution from the shape of genealogical trees. eLife 3:e03568. doi:10.7554/eLife.03568
}
\keyword{LBI}
\keyword{branching}
\keyword{index,}
\keyword{local}

