% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{M}
\alias{M}
\title{M}
\usage{
M(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo} or \code{treeshape}.}
}
\value{
An object of class \code{numeric}.
}
\description{
This calculates M (as defined in Pompei et al., 2012), a measure of asymmetry based on the topological distance, \deqn{M_i}, between tip \deqn{i} and the root:
\deqn{ M = \frac{1}{N} \sum_{i \in \mathcal{L}} {M_i} }.
This is equivalent to Sackin's index divided by the number of tips (also known as the normalised Sackin index).
}
\examples{
N=30
tree<-rtreeshape(1,tip.number=N,model="pda")[[1]]
M(tree)
}
\references{
Pompei S, Loreto V, Tria F (2012) Phylogenetic Properties of RNA Viruses. PLoS ONE 7(9): e44849. doi:10.1371/journal.pone.0044849
Sackin MJ (1972) Good and bad phenograms. Systematic Biology 21: 225-226.
}
\keyword{Sackin's}
\keyword{asymmetry,}
\keyword{imbalance,}
\keyword{index}

