% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matchSimNode.R
\name{matchSimNode}
\alias{matchSimNode}
\title{matchSimNode}
\usage{
matchSimNode(obstree, simtree, node)
}
\arguments{
\item{obstree}{an object of class \code{phylo}}

\item{simtree}{an object of class \code{phylo}}

\item{node}{a number}
}
\value{
the number of the node of height nheight in the simulated tree
}
\description{
A function that takes a node number in an observed tree, and finds the matching node in a tree simulated with \code{getSimTree}.
}
\examples{
tree<-rtree(8)
tree2<-getSimTree(tree)
}

