% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{meanIprime}
\alias{meanIprime}
\title{meanIprime}
\usage{
meanIprime(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo} or \code{treeshape}.}
}
\value{
An object of class \code{numeric}.
}
\description{
This calculates the mean of I' (Purvis and Agapow, 2012) for a phylogenetic tree. This asymmetry metric uses a weighted average of the balance of internal nodes:
\deqn{ meanIprime = \frac{1}{(N-1)} \sum_{j \in \mathcal{I}} w_j \frac{\max(r_j,l_j)-m_j}{r_j+l_j-m_j-1}, } where \eqn{w_j = \frac{r_j+l_j-1}{r_j+l_j}} if \eqn{n} is even and \eqn{w_j = 1} if \eqn{n} is odd.
}
\examples{
N=30
tree<-rtreeshape(1,tip.number=N,model="pda")[[1]]
meanIprime(tree)
}
\references{
\itemize{
 \item Purvis A, Katzourakis A, Agapow PM (2002) Evaluating phylogenetic tree shape: Two modifications to Fusco and Cronk’s method. Journal of Theoretical Biology 214:99–103.
 \item Fusco G, Cronk Q (1995) A new method for evaluating the shape of large phylogenies. Journal of Theoretical Biology 175: 235?243. doi: 10.1006/jtbi.1995.0136
}
}
\keyword{asymmetry}
\keyword{imbalance,}

