% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/imbalanceMetrics.R
\name{Ncherries}
\alias{Ncherries}
\title{Ncherries}
\usage{
Ncherries(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo} or \code{treeshape}.}
}
\value{
An object of class \code{numeric}.
}
\description{
This calculates the number of cherries in a phylogenetic tree, where a cherry represents two tips that share a direct ancestor.
}
\examples{
N=30
tree<-rtreeshape(1,tip.number=N,model="pda")[[1]]
Ncherries(tree)
}
\references{
McKenzie A, Steel M (2000) Distributions of cherries for two models of trees. Math Biosci 164: 81-92.
}
\keyword{asymmetry}
\keyword{imbalance,}

