% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{snt}
\alias{snt}
\title{snt}
\usage{
snt(phy, tm = NULL, nsteps = NULL)
}
\arguments{
\item{tree}{an object of class \code{phylo}.}
}
\value{
A list containing the node times, and Sackin's index at those times.
}
\description{
Gets the Sackin index over the history of a phylogenetic tree.
}
\examples{
tree<-rtree(8)
sackint<-snt(tree)
}

