% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ct}
\alias{ct}
\title{ct}
\usage{
ct(tree)
}
\arguments{
\item{tree}{an object of class \code{phylo} or \code{treeshape}.}
}
\value{
A list containing the node times, and number of cherries at those times.
}
\description{
Gets the number of cherries over the history of a phylogenetic tree.
}
\examples{
tree<-rtree(8)
cherries<-ct(tree)
}

