% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getSimTree}
\alias{getSimTree}
\title{getSimTree}
\usage{
getSimTree(tree)
}
\arguments{
\item{tree}{an object of class \code{phylo}}
}
\value{
a tree of class \code{phylo} with the same tip and internal node times as the input tree.
}
\description{
A function to permute a tree with the same tip and internal node times.
}
\examples{
tree<-rtree(8)
tree2<-getSimTree(tree)
}

