% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{getTipTimes}
\alias{getTipTimes}
\title{getTipTimes}
\usage{
getTipTimes(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo}.}
}
\value{
A vector of times.
}
\description{
Gets the times of the tips of a heterochronously sampled tree with most recently sampled tip set to 0.
}

