% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{I1}
\alias{I1}
\title{I1}
\usage{
I1(tree)
}
\arguments{
\item{tree}{A tree of class \code{phylo} or \code{treeshape}.}
}
\value{
An object of class \code{numeric}.
}
\description{
This calculates I1, a weighted average of the balance of internal nodes, where \eqn{N} is the number of tips, \deqn{j} is the number of nodes, and \eqn{r_j} and \eqn{l_j} represent the number of tips in the left and right subtrees respectively. Then,
\deqn{ I_1 = \frac{2}{(N-1)(N-2)} \sum_{j \in \mathcal{I} } {|r_j-l_j|} }.
I1 is closely related to the Colless index, which can be found using \code{\link[apTreeshape]{colless}}, or by multiplying I1 by \deqn{\frac{(N-1)(N-2)}{2}}.
}
\examples{
N=30
tree<-rtreeshape(1,tip.number=N,model="pda")[[1]]
I1(tree)
}
\references{
\itemise{
 \item Pompei S, Loreto V, Tria F (2012) Phylogenetic Properties of RNA Viruses. PLoS ONE 7(9): e44849. doi:10.1371/journal.pone.0044849
 \item Purvis A, Agapow PM (2002) Phylogeny imbalance: Taxonomic level matters. Systematic Biology 51: 844-854.
 \item Fusco G, Cronk Q (1995) A new method for evaluating the shape of large phylogenies. Journal of Theoretical Biology 175: 235-243. doi: 10.1006/jtbi.1995.0136
}
}
\keyword{Colless}
\keyword{Index}
\keyword{asymmetry,}
\keyword{imbalance,}

